Replace draw() function with display().
The main rendering function has been changed from draw to display. This
was done to differentiate between drawing methods/functions for objects
and the vroom rendering function. Objects which are rendered have draw
methods and there is a draw function which is used to batch render
collections of objects. This syntax is more consistent and hopefully
less confusing going forward.